Use a trait for UiElement implementation
